Add Scene interface
Thought about having the render method take a Renderer class, that will
have stuff like

class Renderer {
  renderPaddle(Paddle paddle)
}

or something similar because then I could abstract away some of the
batch and world units details away from the core logic of the game
itself which could then lend itself to proper unit testing of the scenes
themselves to ensure that the states are updated as expected or similar.

Not 100% sold on the idea to myself yet though, but I do like the idea
of it.